pipeline/requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
openai>=1.50.0
python-dotenv>=1.0.0
requests>=2.31.0
